Seperate Dropzone component
