+++ frontend/src/App.tsx
@@ -1,31 +1,11 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import Button from '@mui/material/Button';
-import { useDropzone } from 'react-dropzone';
-
-function OurFriendlyDropzone() {
-  // @ts-ignore
-  const onDrop = useCallback((acceptedFiles) => {
-    // Do something with the files
-    console.log('acceptedFiles', acceptedFiles);
-  }, []);
-  const { getRootProps, getInputProps, isDragActive } = useDropzone({ onDrop });
-
-  return (
-    <div {...getRootProps()}>
-      <input {...getInputProps()} />
-      {isDragActive ? (
-        <p>Drop the files here ...</p>
-      ) : (
-        <p>Drag 'n' drop some files here, or click to select files</p>
-      )}
-    </div>
-  );
-}
+import Dropzone from './components/Dropzone';
 
 function App() {
   return (
     <>
-      <OurFriendlyDropzone />
+      <Dropzone />
       <Button variant="contained">Hello World</Button>
     </>
   );
